+++ src/simulation/nbody.h
@@ -10,7 +10,7 @@
  * @param region_id     Only performs computation if the particle resides in the particular region.
  *                      To compute for all regions, pass -1 for region_id.
  */
-void update_position(Spec spec, Particle *p, long long n, long double dt, int region_id);
+void update_position(Spec spec, Particle *p, int n, long double dt, int region_id);
 
 /**
  * Computes the new velocity for each particle for a given timestep.
@@ -25,4 +25,4 @@ void update_position(Spec spec, Particle *p, long long n, long double dt, int re
  * @param horizon       Only computes the force by particles within this horizon distance threshold.
  *                      To compute the forces applied by particles in all regions, pass -1 for horizon.
  */
-void update_velocity(Spec spec, Particle *p, long long n, long double dt, int region_id, int horizon);
+void update_velocity(Spec spec, Particle *p, int n, long double dt, int region_id, int horizon);
